log4

+++ src/main/java/com/xq/crowd_funding/raisefunding/beans/pojo/TMemberConfirmInfoPO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class TMemberConfirmInfoPO {
 
-  private Long id;
-  private Long memberid;
+  private Integer id;
+  private Integer memberid;
   private String paynum;
   private String cardnum;
 }
